[ANDROSDK-610] Fix unit tests
