--- core/src/test/java/org/hisp/dhis/android/core/settings/SystemSettingModelBuilderShould.java
@@ -1,62 +1,0 @@
-/*
- * Copyright (c) 2004-2019, University of Oslo
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- *
- * Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- * Neither the name of the HISP project nor the names of its contributors may
- * be used to endorse or promote products derived from this software without
- * specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package org.hisp.dhis.android.core.settings;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.JUnit4;
-
-import java.io.IOException;
-import java.text.ParseException;
-import java.util.List;
-
-import static org.assertj.core.api.Java6Assertions.assertThat;
-
-@RunWith(JUnit4.class)
-public class SystemSettingModelBuilderShould {
-
-    private SystemSetting settingsPojo = SystemSetting.create("aFlag", "aStyle");
-    private SystemSettingModelBuilder modelBuilder = new SystemSettingModelBuilder();
-
-    @Test
-    public void build_flag_setting() throws IOException, ParseException {
-        List<SystemSettingModel> modelList = modelBuilder.splitSettings(settingsPojo);
-        SystemSettingModel flagModel = modelList.get(0);
-        assertThat(flagModel.key()).isEqualTo("flag");
-        assertThat(flagModel.value()).isEqualTo("aFlag");
-    }
-
-    @Test
-    public void build_style_setting() throws IOException, ParseException {
-        List<SystemSettingModel> modelList = modelBuilder.splitSettings(settingsPojo);
-        SystemSettingModel flagModel = modelList.get(1);
-        assertThat(flagModel.key()).isEqualTo("style");
-        assertThat(flagModel.value()).isEqualTo("aStyle");
-    }
-}
